Refactor question/answer counting into its own service
New service -- CompletionService -- keeps track of the number of questions and their answers in a flat array that is quick to analyze.  It provides answered and total question counts.
