+++ CSETWebNg/src/app/assessment/questions/maturity-questions/maturity-questions.component.html
@@ -24,22 +24,23 @@
   <div id="top"></div>
   <div class="d-flex flex-row flex-wrap justify-content-between align-items-start flex-11a mb-4">
 
-    <div class="d-flex flex-column align-items-end flex-00a w-100 mb-3">
-
-      <div class="mt-2">
-        Questions Answered: {{questionsSvc.answerCompletion.answeredCount}}/{{questionsSvc.answerCompletion.totalCount}}
-      </div>
-      
-      <div class="d-flex align-items-start flex-11a">
+    <div class="d-flex flex-row flex-wrap sticky-sal justify-content-end align-items-start flex-11a pb-3">
+      <h6 class="mt-2 mr-3" *ngIf="!!groupings">
+        Answered:
+        {{ completionSvc.answeredCount }}/{{ completionSvc.totalCount }}
+      </h6>
+      <div>
         <button class="btn btn-primary m-0 questions-header-collapse-button-radius" (click)="expandAll(false)"
           matTooltip="Collapse All" aria-label="Collapse All">
           <i class="fas fa-compress-alt"></i></button>
         <button class="btn btn-primary m-0 br-0" (click)="expandAll(true)" matTooltip="Expand All"
           aria-label="Expand All"><i class="fas fa-expand-alt"></i></button>
-        <button class="btn btn-primary m-0 questions-header-filter-button-radius" (click)="showFilterDialog()" matTooltip="Filter">
+        <button class="btn btn-primary m-0 questions-header-filter-button-radius" (click)="showFilterDialog()"
+          matTooltip="Filter">
           <i class="cset-icons-filter" aria-label="Filter"></i>
         </button>
       </div>
+
     </div>
 
     <div class="d-block mb-2">
