+++ CSETWebNg/src/app/assessment/questions/questions.component.html
@@ -24,25 +24,26 @@
   <div id="top"></div>
   <div class="d-flex flex-row flex-wrap sticky-sal justify-content-between align-items-start flex-11a pb-3">
 
-      <div class="btn-group btn-group-toggle" data-toggle="buttons">
-        <label *ngIf="showQuestionsToggle" class="btn btn-app-mode form-check-label"
-          [class.active]="assessSvc.applicationMode == 'Q'" (click)="setMode('Q')"
-          matTooltip="Questions Mode displays simple questions that correspond to standard requirements">
-          <input name="app_mode" class="form-check-input" type="radio" autocomplete="off"
-            [checked]="assessSvc.applicationMode == 'Q'">
-          Questions Mode
-        </label>
-        <label *ngIf="showRequirementsToggle" class="btn btn-app-mode form-check-label"
-          [class.active]="assessSvc.applicationMode == 'R'" (click)="setMode('R')"
-          matTooltip="Requirements Mode displays requirement text verbatim as it appears in the standard">
-          <input name="app_mode" class="form-check-input" type="radio" autocomplete="off"
-            [checked]="assessSvc.applicationMode == 'R'">
-          Requirements Mode
-        </label>
-      </div>
+    <div class="btn-group btn-group-toggle" data-toggle="buttons">
+      <label *ngIf="showQuestionsToggle" class="btn btn-app-mode form-check-label"
+        [class.active]="assessSvc.applicationMode == 'Q'" (click)="setMode('Q')"
+        matTooltip="Questions Mode displays simple questions that correspond to standard requirements">
+        <input name="app_mode" class="form-check-input" type="radio" autocomplete="off"
+          [checked]="assessSvc.applicationMode == 'Q'">
+        Questions Mode
+      </label>
+      <label *ngIf="showRequirementsToggle" class="btn btn-app-mode form-check-label"
+        [class.active]="assessSvc.applicationMode == 'R'" (click)="setMode('R')"
+        matTooltip="Requirements Mode displays requirement text verbatim as it appears in the standard">
+        <input name="app_mode" class="form-check-input" type="radio" autocomplete="off"
+          [checked]="assessSvc.applicationMode == 'R'">
+        Requirements Mode
+      </label>
+    </div>
 
-      <div class="mt-2">
-        Questions Answered: {{questionsSvc.answerCompletion.answeredCount}}/{{questionsSvc.answerCompletion.totalCount}}
+    <div class="d-flex flex-row">
+      <div class="mt-2 mr-3" *ngIf="!!categories">
+        Answered: {{ completionSvc.answeredCount }}/{{ completionSvc.totalCount }}
       </div>
       <div class="d-flex flex-column align-items-start flex-00a">
         <div class="d-flex align-items-start flex-11a">
@@ -63,6 +64,7 @@
           <label for="auto_load_supp" class="checkbox-custom-label" style="margin: 0 5px;">Auto-load Guidance</label>
         </div>
       </div>
+    </div>
   </div>
 
   <h2>Standard Questions</h2>
