+++ CSETWebNg/src/app/models/questions.model.ts
@@ -55,11 +55,6 @@ export interface MaturityQuestionResponse {
     answerOptions: string[];
 }
 
-export interface AnswerCompletion {
-    answeredCount: number;
-    totalCount: number;
-}
-
 export interface MaturityDomainRemarks {
     group_Id: number;
     domainRemark: string;
